requirements.txt
schedule
arxiv
PyYAML
zai-sdk
markdown